fix: SelfieFeedResponse 데이터 형식 수정

+++ src/main/java/com/sixpack/dorundorun/feature/feed/application/FindUserSummaryService.java
@@ -38,7 +38,7 @@ public class FindUserSummaryService {
 		return new SelfieFeedResponse.UserSummary(
 			user.getNickname(),
 			(int)friendCount,
-			totalDistanceMeters.doubleValue(),
+			totalDistanceMeters,
 			(int)selfieCount
 		);
 	}
